Move notifiers to IPC service
